+++ resources/views/hallazgos/index.blade.php
@@ -43,6 +43,7 @@
 
 			En esta secci&oacute;n podr&aacute; crear, modificar y cerrar hallazgos para los distintos datos maestros relevantes a una organizaci&oacute;n. Estos hallazgos puedes estar relacionados a un proceso, subproceso, o a la organizaci&oacute;n misma<br><br>
 			<div id="cargando"><br></div>
+
 			{!!Form::open(['route'=>'hallazgos_lista','method'=>'POST','class'=>'form-horizontal'])!!}
 			<div class="form-group">
 				{!!Form::label('Seleccione organización',null,['class'=>'col-sm-4 control-label'])!!}
@@ -70,7 +71,7 @@
 			
 			</div>
 
-			@if (isset($issues))
+@if (isset($issues))
 
 				@if ($kind == 0)
 					<h4><b>{{ $org }}: Hallazgos de procesos</b></h4>
@@ -88,8 +89,13 @@
 					<h4><b>{{ $org }}: Hallazgos de auditor&iacute;a</b></h4>
 				@endif
 
+		@foreach (Session::get('roles') as $role)
+			@if ($role != 6)
 				{!! link_to_route('create_hallazgo', $title = 'Agregar Hallazgo', $parameters = ['org'=>$org_id,'kind'=>$kind],$attributes = ['class'=>'btn btn-success'])!!}
-				<table id="datatable-2" class="table table-bordered table-striped table-hover table-heading table-datatable" style="font-size:11px">
+			<?php break; ?>
+			@endif
+		@endforeach
+				<table  style="font-size:11px">
 				<thead>
 					@if ($kind == 0)
 						<th>Proceso<label><input type="text" placeholder="Filtrar" /></label></th>
@@ -111,8 +117,13 @@
 					<th>Estado<label><input type="text" placeholder="Filtrar" /></label></th>
 					<th>Fecha final plan<label><input type="text" placeholder="Filtrar" /></label></th>
 					<th>Evidencia</th>
+		@foreach (Session::get('roles') as $role)
+			@if ($role != 6)
 					<th>Editar</th>
 					<th>Eliminar</th>
+				<?php break; ?>
+			@endif
+		@endforeach
 				</thead>
 
 				@foreach ($issues as $issue)
@@ -129,18 +140,26 @@
 							No tiene documentos
 						@else
 							<div style="cursor:hand" id="descargar_{{ $issue['id'] }}" onclick="descargar(2,'{{$issue['evidence'][0]['url'] }}')"><font color="CornflowerBlue"><u>Descargar</u></font></div>
-
+				@foreach (Session::get('roles') as $role)
+					@if ($role != 6)
 							<img src="assets/img/btn_eliminar.png" height="40px" width="40px" onclick="eliminar_ev({{ $issue['id'] }},2)">
-
 							</br>
+					<?php break; ?>
+					@endif
+				@endforeach
 						@endif
+				@foreach (Session::get('roles') as $role)
+					@if ($role != 6)
 						<td>{!! link_to_route('edit_hallazgo', $title = 'Editar', $parameters = ['org'=>$org_id,'id'=>$issue['id']],$attributes = ['class'=>'btn btn-success'])!!}</td>
 						<td>
 						<button class="btn btn-danger" onclick="eliminar({{ $issue['id'] }},'{{ $issue['name'] }}','hallazgo','el hallazgo')">Eliminar</button></td>
+					<?php break; ?>
+					@endif
+				@endforeach
 					</tr>
 				@endforeach
 
-			@endif
+@endif
 
 
 			</div>
